Bank account
System, look like a bank account site with abilities to register your
own account, put and get money from your data using JSP as a view and
servlets as a controllers. Has a connection to database, so no need to
wait for updating information after act of put or get money.